refactor: Market 모듈과 Settlement 모듈 공통 DTO 이동 - #100
