+++ common/src/main/java/com/back/common/dto/cash/request/PayAndHoldRequestDto.java
@@ -1,6 +1,6 @@
-package com.back.common.feign.cash.request;
+package com.back.common.dto.cash.request;
 
-import com.back.common.feign.cash.enums.RelType;
+import com.back.common.dto.cash.enums.RelType;
 
 import java.math.BigDecimal;
