+++ common/src/main/java/com/back/common/dto/settlement/SettlementTargetOrder.java
@@ -1,10 +1,10 @@
-package com.back.settlement.app.event;
+package com.back.common.dto.settlement;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 // Market 모듈 → Settlement 모듈로 전달되는 데이터
-public record SettlementItemRequest(
+public record SettlementTargetOrder(
         Long orderId,
         Long productId,
         Long buyerId,
