+++ market/src/main/java/com/back/market/adapter/in/ApiV1Market.java
@@ -4,8 +4,8 @@ import com.back.common.response.CommonResponse;
 import com.back.market.dto.request.BiddingRequestDto;
 import com.back.market.dto.response.InstantBuyPriceResponseDto;
 import com.back.market.dto.response.InstantSellPriceResponseDto;
-import com.back.common.feign.cash.response.PayAndHoldResponseDto;
-import com.back.common.feign.cash.response.PaymentCancelResponseDto;
+import com.back.common.dto.cash.response.PayAndHoldResponseDto;
+import com.back.common.dto.cash.response.PaymentCancelResponseDto;
 import com.back.security.principal.AuthPrincipal;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
