+++ market/src/main/java/com/back/market/adapter/in/ApiV1MarketController.java
@@ -6,8 +6,8 @@ import com.back.market.app.MarketFacade;
 import com.back.market.dto.request.BiddingRequestDto;
 import com.back.market.dto.response.InstantBuyPriceResponseDto;
 import com.back.market.dto.response.InstantSellPriceResponseDto;
-import com.back.common.feign.cash.response.PayAndHoldResponseDto;
-import com.back.common.feign.cash.response.PaymentCancelResponseDto;
+import com.back.common.dto.cash.response.PayAndHoldResponseDto;
+import com.back.common.dto.cash.response.PaymentCancelResponseDto;
 import com.back.security.principal.AuthPrincipal;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
