+++ market/src/main/java/com/back/market/app/MarketFacade.java
@@ -7,8 +7,8 @@ import com.back.market.app.usecase.RegisterBidUseCase;
 import com.back.market.dto.request.BiddingRequestDto;
 import com.back.market.dto.response.InstantBuyPriceResponseDto;
 import com.back.market.dto.response.InstantSellPriceResponseDto;
-import com.back.common.feign.cash.response.PayAndHoldResponseDto;
-import com.back.common.feign.cash.response.PaymentCancelResponseDto;
+import com.back.common.dto.cash.response.PayAndHoldResponseDto;
+import com.back.common.dto.cash.response.PaymentCancelResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
