+++ market/src/main/java/com/back/market/app/MarketSupport.java
@@ -5,10 +5,10 @@ import com.back.common.code.SuccessCode;
 import com.back.common.exception.BadRequestException;
 import com.back.common.response.CommonResponse;
 import com.back.market.adapter.out.cash.CashClient;
-import com.back.common.feign.cash.request.PayAndHoldRequestDto;
-import com.back.common.feign.cash.request.PaymentCancelRequestDto;
-import com.back.common.feign.cash.response.PayAndHoldResponseDto;
-import com.back.common.feign.cash.response.PaymentCancelResponseDto;
+import com.back.common.dto.cash.request.PayAndHoldRequestDto;
+import com.back.common.dto.cash.request.PaymentCancelRequestDto;
+import com.back.common.dto.cash.response.PayAndHoldResponseDto;
+import com.back.common.dto.cash.response.PaymentCancelResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
