+++ market/src/main/java/com/back/market/app/usecase/CancelBidUseCase.java
@@ -7,9 +7,9 @@ import com.back.market.app.MarketSupport;
 import com.back.market.domain.Bidding;
 import com.back.market.domain.enums.BiddingPosition;
 import com.back.market.domain.enums.BiddingStatus;
-import com.back.common.feign.cash.enums.RelType;
-import com.back.common.feign.cash.request.PaymentCancelRequestDto;
-import com.back.common.feign.cash.response.PaymentCancelResponseDto;
+import com.back.common.dto.cash.enums.RelType;
+import com.back.common.dto.cash.request.PaymentCancelRequestDto;
+import com.back.common.dto.cash.response.PaymentCancelResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
