+++ market/src/main/java/com/back/market/app/usecase/ConfirmPaymentUseCase.java
@@ -8,8 +8,8 @@ import com.back.market.domain.Bidding;
 import com.back.market.domain.Order;
 import com.back.market.domain.enums.BiddingStatus;
 import com.back.market.domain.enums.OrderStatus;
-import com.back.common.feign.cash.enums.RelType;
-import com.back.common.feign.cash.request.PaymentCompletedRequestDto;
+import com.back.common.dto.cash.enums.RelType;
+import com.back.common.dto.cash.request.PaymentCompletedRequestDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
