+++ market/src/main/java/com/back/market/app/usecase/RegisterBidUseCase.java
@@ -11,11 +11,11 @@ import com.back.market.domain.MarketProduct;
 import com.back.market.domain.MarketUser;
 import com.back.market.domain.enums.BiddingPosition;
 import com.back.market.domain.enums.BiddingStatus;
-import com.back.common.feign.cash.enums.PayAndHoldStatus;
-import com.back.common.feign.cash.enums.RelType;
+import com.back.common.dto.cash.enums.PayAndHoldStatus;
+import com.back.common.dto.cash.enums.RelType;
 import com.back.market.dto.request.BiddingRequestDto;
-import com.back.common.feign.cash.request.PayAndHoldRequestDto;
-import com.back.common.feign.cash.response.PayAndHoldResponseDto;
+import com.back.common.dto.cash.request.PayAndHoldRequestDto;
+import com.back.common.dto.cash.response.PayAndHoldResponseDto;
 import com.back.market.mapper.BiddingMapper;
 import com.back.market.mapper.CashRequestMapper;
 import lombok.RequiredArgsConstructor;
