+++ market/src/test/java/com/back/market/MatchInstantTradeUseCaseV2Tests.java
@@ -15,9 +15,9 @@ import com.back.market.domain.enums.BiddingPosition;
 import com.back.market.domain.enums.BiddingStatus;
 import com.back.market.domain.enums.OrderStatus;
 import com.back.market.domain.enums.Role;
-import com.back.common.feign.cash.enums.PayAndHoldStatus;
+import com.back.common.dto.cash.enums.PayAndHoldStatus;
 import com.back.market.dto.request.BiddingRequestDto;
-import com.back.common.feign.cash.response.PayAndHoldResponseDto;
+import com.back.common.dto.cash.response.PayAndHoldResponseDto;
 import com.back.market.mapper.BiddingMapper;
 import com.back.market.mapper.MarketProductMapper;
 import com.back.market.mapper.MarketUserMapper;
