+++ market/src/test/java/com/back/market/RegisterBidUseCaseV2Tests.java
@@ -12,9 +12,9 @@ import com.back.market.domain.MarketUser;
 import com.back.market.domain.enums.BiddingPosition;
 import com.back.market.domain.enums.BiddingStatus;
 import com.back.market.domain.enums.Role;
-import com.back.common.feign.cash.enums.PayAndHoldStatus;
+import com.back.common.dto.cash.enums.PayAndHoldStatus;
 import com.back.market.dto.request.BiddingRequestDto;
-import com.back.common.feign.cash.response.PayAndHoldResponseDto;
+import com.back.common.dto.cash.response.PayAndHoldResponseDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
